Finishing first stage of messaging system projects
After managing to adjust the chats, within the new parent container, the idea arise: what happens if I have more chats than the viewport can display? To do this, I applied overflow on the Y axis to generate automatic scrolling, this challenge forced me to perfectly understand how flex-grow, flex-shrink and flex-basis work and which one is the way to apply it. Next step, use media queriers to create a mobile version